Move label finding code to FunctionQParameters file
